Test binomial on MiteMap for presence in the half part of the odor

The test is run for each Modality values and each Farm with a p-adjustement step.

Usage

binom_test_mitemap(MiteMap, modality = "Modality", p.adjust_method = "BH")

Arguments

MiteMap

(required) The result of import_mitemap ($resulting_data) for raw_data

modality

The modality to separate test

p.adjust_method

method for p-adjustement. See p.adjust()

Value

A tibble of results for binomial test

Author

Adrien Taudière

Examples

binom_test_mitemap(HH_example$resulting_data)
#> # A tibble: 6 × 10
#> # Groups:   Farm [2]
#>   Farm     Modality     n    HH   yes    no p.value p.value.adj estimate CI     
#>   <chr>    <fct>    <int> <dbl> <dbl> <dbl>   <dbl>       <dbl>    <dbl> <chr>  
#> 1 POUL6_D1 G           31 0.129     4    27   0.001       0.003    0.129 0.036 …
#> 2 POUL6_D1 NM          31 0.129     4    27   0.001       0.003    0.129 0.036 …
#> 3 POUL6_D1 T           31 0.387    12    19   0.281       0.337    0.387 0.218 …
#> 4 POUL6_P1 G           29 0.207     6    23   0.002       0.004    0.207 0.08 -…
#> 5 POUL6_P1 NM          28 0.214     6    22   0.004       0.006    0.214 0.083 …
#> 6 POUL6_P1 T           27 0.519    14    13   1           1        0.519 0.319 …
binom_test_mitemap(HH_example$resulting_data, p.adjust_method = "bonferroni")
#> # A tibble: 6 × 10
#> # Groups:   Farm [2]
#>   Farm     Modality     n    HH   yes    no p.value p.value.adj estimate CI     
#>   <chr>    <fct>    <int> <dbl> <dbl> <dbl>   <dbl>       <dbl>    <dbl> <chr>  
#> 1 POUL6_D1 G           31 0.129     4    27   0.001       0.006    0.129 0.036 …
#> 2 POUL6_D1 NM          31 0.129     4    27   0.001       0.006    0.129 0.036 …
#> 3 POUL6_D1 T           31 0.387    12    19   0.281       1        0.387 0.218 …
#> 4 POUL6_P1 G           29 0.207     6    23   0.002       0.012    0.207 0.08 -…
#> 5 POUL6_P1 NM          28 0.214     6    22   0.004       0.024    0.214 0.083 …
#> 6 POUL6_P1 T           27 0.519    14    13   1           1        0.519 0.319 …