Filter MiteMap result

Filter MiteMap result

Usage

filter_mitemap(
  MiteMap,
  HH = NULL,
  CH = NULL,
  first_seconds_to_delete = 2,
  bad_range_value_x = 50,
  bad_range_value_y = 50,
  max_x_value = 42,
  min_x_value = 0,
  max_y_value = 42,
  min_y_value = 0,
  maximum_time = 603,
  center_x = 0,
  center_y = 0
)

Arguments

MiteMap

(required) The result of import_mitemap ($resulting_data) for raw_data

HH

The result of import_mitemap() ($resulting_data) for type_of_files = "HH". It remove run with (i) time in <0, time out <0 and time_in/(time_in + time_out) > 1 in the HH files.

CH

The result of import_mitemap() ($resulting_data) for type_of_files = "CH". It remove run with (i) time in <0, time out <0 and time_in/(time_in + time_out) > 1 in the CH files.

first_seconds_to_delete

(default: 2) How many seconds we need to delete?

bad_range_value_x

(default: 42) The range value of x that is superior to what we expect. This param filter values on the File_name basis.

bad_range_value_y

(default: 42) The range value of y that is superior to what we expect. This param filter values on the File_name basis.

max_x_value

(default: 42) Maximum value for x axis. This param filter values on the point (one line) basis.

min_x_value

(default: 0) Minimum value for x axis. This param filter values on the point (one line) basis.

max_y_value

(default: 42) Maximum value for y axis. This param filter values on the point (one line) basis.

min_y_value

(default: 0) Minimum value for y axis. This param filter values on the point (one line) basis.

maximum_time

(default: 603) Maximum time for one run (in seconds)).

center_x

(int, default 0) Center the value of x by additioning center_x mm to x.mm.

center_y

(int, default 0) Center the value of y by additioning center_y mm to y.mm.

Value

a filtered resulting table of class tibble

Examples

MM_filtered_centered <- filter_mitemap(MM_example,
  center_x = -20, center_y = -20,
  bad_range_value_x = 50, bad_range_value_y = 50
)
#> Row removed when clearing the first secondes: 536
#> Row removed when clearing bad x range: 16860
#> Row removed when clearing bad y range: 0
#> Row removed when clearing bad x values: 24537
#> Row removed when clearing bad y values: 93
#> Row removed when clearing for run with times sup to maximum_time: 814
filter_mitemap(MM_example,
  HH = HH_example$resulting_data,
  bad_range_value_x = 50, bad_range_value_y = 50
)
#> Row removed when clearing the first secondes: 536
#> Row removed when clearing bad x range: 16860
#> Row removed when clearing bad y range: 0
#> Row removed when clearing bad x values: 24537
#> Row removed when clearing bad y values: 93
#> Row removed when clearing for bad HH and/or CH: 814
#> Row removed when clearing for run with times sup to maximum_time814
#> # A tibble: 78,785 × 12
#> # Groups:   File_name [35]
#>    File_name X..t.s. x.mm. y.mm. Immobile..si..1. Run_number Date             
#>    <chr>       <dbl> <dbl> <dbl>            <dbl>      <int> <dttm>           
#>  1 MM062022…    17.8  41.9  31.9                0         20 22-03-19 00:00:00
#>  2 MM062022…    18    41.8  32.1                0         20 22-03-19 00:00:00
#>  3 MM062022…    18.2  41.5  32.3                0         20 22-03-19 00:00:00
#>  4 MM062022…    18.4  41.4  32.5                0         20 22-03-19 00:00:00
#>  5 MM062022…    18.6  41.3  32.7                0         20 22-03-19 00:00:00
#>  6 MM062022…    18.8  41    33                  0         20 22-03-19 00:00:00
#>  7 MM062022…    19    40.9  33.3                0         20 22-03-19 00:00:00
#>  8 MM062022…    19.2  40.6  33.5                0         20 22-03-19 00:00:00
#>  9 MM062022…    19.4  40.5  33.6                0         20 22-03-19 00:00:00
#> 10 MM062022…    19.6  40.2  33.8                0         20 22-03-19 00:00:00
#> # ℹ 78,775 more rows
#> # ℹ 5 more variables: Start_time <chr>, Farm <chr>, MiteMap_number <chr>,
#> #   Bag <int>, Modality <chr>