Import MiteMapTools data

Import MiteMapTools data

Usage

import_mitemap(
  path_to_folder,
  path_to_metadata = NULL,
  format_metadata = "csv",
  type_of_files = "raw_data",
  delete_parenthesis = FALSE,
  replace_parenthesis = TRUE,
  delete_space = TRUE,
  messages = TRUE,
  csv_with_correction = FALSE,
  remove_csv_folder = TRUE,
  dec = ",",
  force = FALSE,
  skip = 1,
  colnames_metadata = c("Run_number", "File_name", "Date", "Start_time", "Farm",
    "MiteMap_number", "Bag", "Modality")
)

Arguments

path_to_folder

Path to the folder with zip files containing csv and with one metadata files in the .xlsx format

path_to_metadata

Path to the csv file (or xlsx if format_metadata="xlsx"). If there is only one csv file set path_to_metadata="csv" (or "xlsx"). Need to be set using a complete path to csv file if this file is not in the folder or if their is multiple xlsx/csv files in the folder.

format_metadata

Either csv or xlsx

type_of_files

(default : "raw_data") A pattern to select the type of csv (either "raw_data", "CH" or "HH") files inside the zip files.

delete_parenthesis

(Logical, default FALSE) Do we delete parenthesis with a number inside in the name of the files. Note that the name of the csv inside a zip file with a parenthesis do not have parenthesis into this name. Thus the default value to TRUE is recommended either in delete_parenthesis or in replace_parenthesis

replace_parenthesis

(Logical, default TRUE) Replace xxx(1) by xxx_1

delete_space

(Logical, default TRUE)

messages

(Logical, default TRUE) Do we print some warnings?

csv_with_correction

(Logical, default FALSE) If TRUE an if present in the zip files, position files with correction (center and reduce to 0) are used.

remove_csv_folder

(Logical, default TRUE) If FALSE, the csv_folder is kept. May be useful for debugging.

dec

decimal for the csv metadata files.

force

Force overwriting the path to csv_folder

skip

(default 1) skip the first line of the metadata files.

colnames_metadata

Column names for metadata column. Note that File_name is required and must correspond to the names of zip files.

Value

A list of 4 elements

resulting_data: a (possibly huge) tibble with metadata information and position in x, y and time.

files_not_in_csv: a list of files not present in csv filenames from zip files.

files_not_in_metada: a list of files not funded in metadata.

duplicate_file_name_in_metadata: a list of duplicated file names in metadata.

By default, the resulting_data slot is structured in 12 columns (The fifth ones are from zip files and the other from metadata files).

File_name

The time in second (position is recorded every 0.2s)

The position in x (in mm)

The position in y (in mm)

Boolean variable indicating if the individual has remained immobile since the last record (1 if immobile)

Run number

Date

Start time

Farm

MiteMap number

Bag

Modality

Author

Adrien Taudière

Examples

mm_csv <- suppressWarnings(import_mitemap(
  system.file("extdata", "POUL6", package = "MiteMapTools")
))
#> 
#> The final number of samples for folder is 117721.
dim(mm_csv$resulting_data)
#> [1] 117721     12

mm_HH <- suppressWarnings(import_mitemap(
  system.file("extdata", "POUL6", package = "MiteMapTools"),
  type_of_files = "HH"
))
#> 
#> The final number of samples for folder is 43.
dim(mm_HH$resulting_data)
#> [1] 43 16

mm_xlsx <- suppressWarnings(import_mitemap(
  system.file("extdata", "POUL6", package = "MiteMapTools"),
  format_metadata = "xlsx"
))
#> 
#> The final number of samples for folder is 117721.

dim(mm_xlsx$resulting_data)
#> [1] 117721     12