Create folder(s) with heatmap

Create folder(s) with heatmap

Usage

make_heatmap(
  path_to_folder,
  output_path = "Heatmap",
  path_to_metadata = NULL,
  format_metadata = "csv",
  modality = "Modality",
  subfolder_by_farm = TRUE,
  subfolder_by_modality = TRUE,
  dec = ",",
  skip = 1,
  colnames_metadata = c("Run_number", "File_name", "Date", "Start_time", "Farm",
    "MiteMap_number", "Bag", "Modality"),
  force = FALSE
)

Arguments

path_to_folder

Path to folder

output_path

(default = "Heatmap")

path_to_metadata

Path to metadata

format_metadata

(either csv of xlsx)

modality

(character, default "Modality") Set to a column names to create subfolder for each levels of the modality

subfolder_by_farm

(Logical, default = TRUE) create subfolder for each farm

subfolder_by_modality

(Logical, default = TRUE) create subfolder for each modality set in parameter modality

dec

decimal for the csv metadata files.

skip

(default 1) skip the first line of the metadata files.

colnames_metadata

Column names for metadata column. Note that File_name is required and must correspond to the names of zip files.

force

Force overwriting the path to output_path

Value

Nothing. Create files in folders.

Author

Adrien Taudière

Examples

make_heatmap(system.file("extdata", "POUL6", package = "MiteMapTools"))
unlink("Heatmap", recursive = TRUE)