Import MiteMapTools data

This function import data from zip files generated by MiteMap software. It unzips all zip files in a folder, read the csv files and merge them with metadata if provided. It also compute some metrics such as speed and turning angle. By default, it also clean the data using filter_mitemap() with default parameters. See ?filter_mitemap() for other parameters or set clean to FALSE if you want raw data. By default the center of the arena is located at (0,0) with a diameter of 40mm and the source of the odor is located at c(-20 , 0).

If your zip files or metadata files have parenthesis and space in the name, you may want to use rename_files_with_number() function before using this function. The file name in the metadata must match the zip file name, but import_mitemap() integrate some options to handle parenthesis and space in the name of the files in the metadata.

Usage

import_mitemap(
  path_to_folder,
  path_to_metadata = NULL,
  format_metadata = "csv",
  clean = TRUE,
  delete_parenthesis = FALSE,
  replace_parenthesis = TRUE,
  delete_space = TRUE,
  csv_with_correction = FALSE,
  remove_csv_folder = TRUE,
  dec = ",",
  sep = ";",
  n_lag = 1,
  center_x = 0,
  center_y = 0,
  x_odor = -20,
  y_odor = 0,
  force = FALSE,
  return_with_logs = FALSE,
  compute_metrics = TRUE,
  file_name_column = "File_name",
  raw_data_name = "donnees_brutes",
  radius_CH = 23.175,
  verbose = TRUE,
  ...
)

Arguments

path_to_folder

Path to the folder with zip files containing csv and with one metadata files in the .xlsx format

path_to_metadata

Path to the csv file (or xlsx if format_metadata="xlsx"). If there is only one csv file set path_to_metadata="csv" (or "xlsx"). Need to be set using a complete path to csv file if this file is not in the folder or if their is multiple xlsx/csv files in the folder.

format_metadata

Either csv or xlsx. If set to NULL, no metadata is used.

clean

(logical, default TRUE) Do we clean the MiteMap result using filter_mitemap() function? See filter_mitemap() for additional parameters.

delete_parenthesis

(Logical, default FALSE) Do we delete parenthesis with a number inside in the name of the files in the metadata. Note that the name of the csv inside a zip file with a parenthesis do not have parenthesis into this name. Thus, we recommended to set TRUE at least in one of delete_parenthesis or replace_parenthesis parameter.

replace_parenthesis

(Logical, default TRUE) Replace abc_name(1) by abc_name_1 in the name of the files in the metadata.

delete_space

(Logical, default TRUE) Delete_space in the name of the files in the metadata.

csv_with_correction

(Logical, default FALSE) If TRUE an if present in the zip files, position files with correction (center and reduce to 0) are used.

remove_csv_folder

(Logical, default TRUE) If FALSE, the csv_folder is kept. May be useful for debugging.

dec

decimal for the csv metadata files.

sep

separator for the csv metadata files.

n_lag

(int, default 1) Number of time step to compute speed and turning angle. If n_lag=1 (default), speed and turning angle are computed between two consecutive time step (0.2s). If n_lag=2 speed and turning angle are computed between two time step before and two time step after.

center_x

(int, default 0) Center the value of x by additioning center_x mm to x.mm.

center_y

(int, default 0) Center the value of y by additioning center_y mm to y.mm.

x_odor

(int, default -20) The x position of the odor source.

y_odor

(int, default 0) The y position of the odor source.

force

Force overwriting the path to csv_folder

return_with_logs

(Logical, default FALSE). If TRUE, the returning object is a list of 4 elements containing useful information to explore unmatching name between file_names and metadata

compute_metrics

(Logical, default TRUE). Are metrics such as time_immobile, speed and turning angles are computed for each time step ?

file_name_column

Name for the column corresponding to the File_name.

raw_data_name

Name of the raw data file in the zip file. Default is "donnees_brutes" (sorry for the french name). If using code from https://github.com/LR69/MiteMap/tree/MiteMap.v6 no need to change this.

radius_CH

the radius of the circular half arena (in mm). Default is 23.175 to fit with the area of HH shape.

verbose

(logical, default TRUE). If TRUE, print additional information.

...

Other params for be passed on to filter_mitemap().

Value

If return_with_logs is FALSE (default), the return object is only the tibble called resulting_data. If return_with_logs is TRUE, it return a list of 4 elements:

resulting_data: a (possibly huge) tibble with metadata information and position in x, y and time.

files_not_in_csv: a list of files not present in csv filenames from zip files.

files_not_in_metada: a list of files not funded in metadata.

duplicate_file_name_in_metadata: a list of duplicated file names in metadata.

By default, the resulting_data slot is structured in 4 obligated columns + columns from the metadata file + the computed metrics if compute_metrics is TRUE.

File_name

X..t.s. - The time in second (position is recorded every 0.2s)

x.mm. - The position in x (in mm)

y.mm. - The position in y (in mm)

Metadata columns

Computed metrics

distance - distance between two consecutive time step (in mm)

speed_mm_s - speed between two consecutive time step (in mm/s)

Author

Adrien Taudière

Examples

mm_csv <- import_mitemap(
  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  file_name_column = "File (mite ID)", verbose = FALSE
)
dim(mm_csv)
#> [1] 70828    35
mm_csv$File_name |> unique() |> length()
#> [1] 50

mm_csv2 <- import_mitemap(
  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  file_name_column = "File (mite ID)", verbose = FALSE, return_with_logs = TRUE
)
mm_csv2$files_not_in_csv
#>    [1] "MM032020_10_24_22h23m12s"   "MM042022_09_17_01h23m57s"  
#>    [3] "MM052022_09_16_21h23m45s"   "MM082021_07_08_14h23m09s"  
#>    [5] "MM092022_09_17_14h23m27s"   "MM032020_10_24_22h22m48s"  
#>    [7] "MM042022_09_17_01h24m29s"   "MM052022_09_16_21h23m07s"  
#>    [9] "MM082021_07_08_14h35m25s"   "MM092022_09_17_14h23m05s"  
#>   [11] "MM032020_10_24_22h22m57s"   "MM042022_09_17_01h22m58s"  
#>   [13] "MM052022_09_16_21h23m47s"   "MM082021_07_08_14h22m49s"  
#>   [15] "MM092022_09_17_14h23m03s"   "MM032020_10_24_22h23m01s"  
#>   [17] "MM042022_09_17_01h23m20s"   "MM052022_09_16_21h24m09s"  
#>   [19] "MM082021_07_08_14h23m37s"   "MM092022_09_17_14h23m48s"  
#>   [21] "MM032020_10_24_22h23m52s"   "MM042022_09_17_01h22m49s"  
#>   [23] "MM052022_09_16_21h23m24s"   "MM082021_07_08_14h24m25s"  
#>   [25] "MM092022_09_17_14h24m45s"   "MM032020_10_24_22h23m33s"  
#>   [27] "MM042022_09_17_01h23m33s"   "MM052022_09_16_21h23m47s_1"
#>   [29] "MM082021_07_08_14h23m24s"   "MM092022_09_17_14h23m04s"  
#>   [31] "MM032020_10_24_22h23m27s"   "MM042022_09_17_01h24m00s"  
#>   [33] "MM052022_09_16_21h24m43s"   "MM082021_07_08_15h23m07s"  
#>   [35] "MM092022_09_17_15h23m33s"   "MM032020_10_24_22h23m03s"  
#>   [37] "MM042022_09_17_01h23m32s"   "MM052022_09_16_21h23m45s_1"
#>   [39] "MM082021_07_08_15h23m03s"   "MM092022_09_17_15h23m27s"  
#>   [41] "MM032020_10_24_22h24m03s"   "MM042022_09_17_01h25m44s"  
#>   [43] "MM052022_09_16_21h24m00s"   "MM082021_07_08_16h23m19s"  
#>   [45] "MM032020_10_24_22h23m33s_1" "MM042022_09_17_01h23m29s"  
#>   [47] "MM052022_09_16_21h23m40s"   "MM082021_07_08_16h22m53s"  
#>   [49] "MM032020_10_24_23h23m33s"   "MM042022_09_17_02h23m15s"  
#>   [51] "MM052022_09_16_22h23m09s"   "MM082021_07_08_16h23m39s"  
#>   [53] "MM092022_09_17_16h23m21s"   "MM032020_10_24_23h23m11s"  
#>   [55] "MM042022_09_17_02h23m03s"   "MM052022_09_16_22h23m03s"  
#>   [57] "MM082021_07_08_16h23m11s"   "MM092022_09_17_16h23m02s"  
#>   [59] "MM032020_10_24_23h24m29s"   "MM042022_09_17_02h24m48s"  
#>   [61] "MM052022_09_16_22h23m09s_1" "MM082021_07_08_16h23m59s"  
#>   [63] "MM092022_09_17_16h24m21s"   "MM032020_10_24_23h23m18s"  
#>   [65] "MM042022_09_17_02h23m33s"   "MM052022_09_16_22h23m43s"  
#>   [67] "MM082021_07_08_16h25m01s_1" "MM092022_09_17_16h24m44s"  
#>   [69] "MM032020_10_24_23h23m21s"   "MM042022_09_17_02h22m54s"  
#>   [71] "MM052022_09_16_22h23m12s"   "MM082021_07_08_16h23m56s"  
#>   [73] "MM092022_09_17_16h24m11s"   "MM032020_10_24_23h36m57s"  
#>   [75] "MM042022_09_17_02h23m05s"   "MM052022_09_16_22h23m03s_1"
#>   [77] "MM082021_07_08_16h23m15s"   "MM092022_09_17_16h22m54s"  
#>   [79] "MM032020_10_24_23h23m09s"   "MM042022_09_17_02h22m51s"  
#>   [81] "MM052022_09_16_22h22m47s"   "MM082021_07_08_16h22m57s"  
#>   [83] "MM092022_09_17_16h23m13s"   "MM032020_10_24_23h22m57s"  
#>   [85] "MM042022_09_17_02h23m43s"   "MM052022_09_16_22h22m57s"  
#>   [87] "MM082021_07_08_16h23m28s"   "MM092022_09_17_16h24m23s"  
#>   [89] "MM032020_10_24_23h22m55s"   "MM042022_09_17_02h23m13s"  
#>   [91] "MM052022_09_16_22h23m23s"   "MM082021_07_08_16h23m03s"  
#>   [93] "MM092022_09_17_16h22m55s"   "MM012022_05_17_10h23m27s_1"
#>   [95] "MM032020_10_24_23h23m11s_1" "MM042022_09_17_02h23m19s"  
#>   [97] "MM052022_09_16_22h23m45s"   "MM082021_07_08_16h23m31s"  
#>   [99] "MM092022_09_17_16h24m49s"   "MM032020_10_24_23h23m34s"  
#>  [101] "MM042022_09_17_02h23m22s"   "MM052022_09_16_22h23m01s"  
#>  [103] "MM082021_07_08_16h23m13s"   "MM092022_09_17_16h23m19s"  
#>  [105] "MM032020_10_24_23h24m42s"   "MM042022_09_17_02h22m59s"  
#>  [107] "MM052022_09_16_22h23m55s"   "MM082021_07_08_16h23m25s"  
#>  [109] "MM092022_09_17_16h24m12s"   "MM042022_09_17_02h23m29s"  
#>  [111] "MM052022_09_16_22h24m15s"   "MM082021_07_08_16h22m55s"  
#>  [113] "MM092022_09_17_16h23m07s"   "MM032020_10_24_23h23m12s"  
#>  [115] "MM042022_09_17_02h23m15s_1" "MM052022_09_16_22h23m10s"  
#>  [117] "MM082021_07_08_16h22m53s_1" "MM032020_10_24_23h23m19s"  
#>  [119] "MM042022_09_17_02h23m15s_2" "MM052022_09_16_22h23m13s"  
#>  [121] "MM082021_07_08_16h23m25s_1" "MM092022_09_17_16h23m45s"  
#>  [123] "MM032020_10_24_23h23m12s_1" "MM042022_09_17_02h23m11s"  
#>  [125] "MM052022_09_16_22h23m51s"   "MM082021_07_08_16h23m22s"  
#>  [127] "MM092022_09_17_16h23m26s"   "MM032020_10_24_23h23m05s"  
#>  [129] "MM042022_09_17_02h22m57s"   "MM052022_09_16_22h23m27s"  
#>  [131] "MM082021_07_08_16h22m59s"   "MM092022_09_17_16h22m50s"  
#>  [133] "MM032020_10_24_23h23m52s"   "MM042022_09_17_02h23m47s"  
#>  [135] "MM052022_09_16_22h24m19s"   "MM082021_07_08_16h23m03s_1"
#>  [137] "MM092022_09_17_16h23m51s"   "MM032020_10_24_23h23m18s_1"
#>  [139] "MM042022_09_17_02h23m27s"   "MM052022_09_16_22h23m39s"  
#>  [141] "MM082021_07_08_16h22m52s"   "MM092022_09_17_16h23m11s_1"
#>  [143] "MM042022_09_17_02h23m33s_1" "MM052022_09_16_22h23m39s_1"
#>  [145] "MM082021_07_08_16h23m17s"   "MM092022_09_17_16h23m38s"  
#>  [147] "MM042022_09_17_02h22m46s"   "MM052022_09_16_22h22m49s"  
#>  [149] "MM082021_07_08_16h23m11s_1" "MM032020_10_24_23h22m50s"  
#>  [151] "MM042022_09_17_02h23m22s_1" "MM052022_09_16_22h23m21s"  
#>  [153] "MM082021_07_08_16h23m53s"   "MM092022_09_17_16h24m03s"  
#>  [155] "MM032020_10_24_23h23m51s"   "MM042022_09_17_02h23m51s"  
#>  [157] "MM052022_09_16_22h23m28s"   "MM082021_07_08_16h23m35s"  
#>  [159] "MM092022_09_17_16h23m21s_1" "MM032020_10_24_23h23m06s"  
#>  [161] "MM042022_09_17_02h23m15s_3" "MM052022_09_16_22h23m19s"  
#>  [163] "MM082021_07_08_16h23m37s"   "MM092022_09_17_16h23m32s"  
#>  [165] "MM032020_10_24_23h23m32s"   "MM042022_09_17_02h23m43s_1"
#>  [167] "MM052022_09_16_22h23m49s"   "MM082021_07_08_16h24m19s"  
#>  [169] "MM092022_09_17_16h23m56s"   "MM032020_10_24_23h23m41s"  
#>  [171] "MM042022_09_17_02h22m51s_1" "MM052022_09_16_22h23m09s_2"
#>  [173] "MM082021_07_08_16h24m13s"   "MM092022_09_17_16h23m55s"  
#>  [175] "MM012022_05_17_10h24m31s_1" "MM032020_10_24_23h23m04s"  
#>  [177] "MM042022_09_17_02h23m48s"   "MM052022_09_16_22h24m16s"  
#>  [179] "MM082021_07_08_16h22m59s_1" "MM092022_09_17_16h22m47s"  
#>  [181] "MM032020_10_24_23h24m48s"   "MM042022_09_17_02h24m51s"  
#>  [183] "MM052022_09_16_22h25m44s"   "MM082021_07_08_16h25m02s"  
#>  [185] "MM092022_09_17_16h26m18s"   "MM032020_10_24_23h22m57s_1"
#>  [187] "MM042022_09_17_02h23m33s_2" "MM052022_09_16_22h23m30s"  
#>  [189] "MM082021_07_08_16h23m29s"   "MM092022_09_17_16h23m59s"  
#>  [191] "MM032020_10_25_00h23m26s"   "MM042022_09_17_03h23m03s"  
#>  [193] "MM052022_09_16_23h22m47s"   "MM082021_07_08_17h23m59s"  
#>  [195] "MM092022_09_17_17h23m41s"   "MM032020_10_25_00h24m02s"  
#>  [197] "MM042022_09_17_03h23m46s"   "MM052022_09_16_23h23m47s"  
#>  [199] "MM082021_07_08_17h22m51s"   "MM032020_10_25_00h24m04s"  
#>  [201] "MM042022_09_17_03h23m19s"   "MM052022_09_16_23h24m05s"  
#>  [203] "MM082021_07_08_17h23m05s"   "MM092022_09_17_17h23m06s"  
#>  [205] "MM012022_05_17_11h24m01s_1" "MM032020_10_25_00h29m16s"  
#>  [207] "MM042022_09_17_03h24m40s"   "MM052022_09_16_23h25m13s"  
#>  [209] "MM082021_07_08_17h24m51s"   "MM092022_09_17_17h25m06s"  
#>  [211] "MM032020_10_25_00h22m55s"   "MM042022_09_17_03h22m53s"  
#>  [213] "MM052022_09_16_23h23m17s"   "MM082021_07_08_17h23m51s"  
#>  [215] "MM092022_09_17_17h24m25s"   "MM032020_10_25_00h24m00s"  
#>  [217] "MM042022_09_17_03h22m54s"   "MM052022_09_16_23h24m48s"  
#>  [219] "MM082021_07_08_17h23m01s"   "MM092022_09_17_17h23m33s"  
#>  [221] "MM032020_10_25_00h24m16s"   "MM042022_09_17_03h24m09s"  
#>  [223] "MM052022_09_16_23h24m20s"   "MM082021_07_08_17h24m18s"  
#>  [225] "MM092022_09_17_17h24m42s"   "MM032020_10_25_00h23m16s"  
#>  [227] "MM042022_09_17_03h23m16s"   "MM052022_09_16_23h24m09s"  
#>  [229] "MM082021_07_08_17h24m57s"   "MM092022_09_17_17h24m34s"  
#>  [231] "MM082021_07_08_17h23m03s"   "MM092022_09_17_17h24m21s"  
#>  [233] "MM012022_05_17_11h22m49s"   "MM082021_07_08_17h22m53s"  
#>  [235] "MM092022_09_17_17h23m52s"   "MM012022_05_17_11h23m33s"  
#>  [237] "MM032020_10_25_02h23m46s"   "MM042022_09_17_05h25m06s"  
#>  [239] "MM052022_09_17_01h23m55s"   "MM082021_07_08_19h24m44s"  
#>  [241] "MM092022_09_17_19h25m09s"   "MM032020_10_25_02h23m33s"  
#>  [243] "MM042022_09_17_05h23m27s"   "MM052022_09_17_01h22m57s"  
#>  [245] "MM082021_07_08_19h23m34s"   "MM092022_09_17_19h23m58s"  
#>  [247] "MM042022_09_17_05h24m25s"   "MM052022_09_17_01h24m23s"  
#>  [249] "MM082021_07_08_19h24m33s"   "MM032020_10_25_02h23m11s"  
#>  [251] "MM042022_09_17_05h23m13s"   "MM052022_09_17_01h23m25s"  
#>  [253] "MM082021_07_08_19h23m13s"   "MM092022_09_17_19h23m10s"  
#>  [255] "MM032020_10_25_02h23m32s"   "MM042022_09_17_05h23m41s"  
#>  [257] "MM052022_09_17_01h24m09s"   "MM082021_07_08_19h23m23s"  
#>  [259] "MM092022_09_17_19h23m54s"   "MM032020_10_25_02h23m03s"  
#>  [261] "MM042022_09_17_05h23m08s"   "MM052022_09_17_01h23m27s"  
#>  [263] "MM092022_09_17_19h23m47s"   "MM032020_10_25_02h25m06s"  
#>  [265] "MM042022_09_17_05h23m55s"   "MM052022_09_17_01h24m17s"  
#>  [267] "MM092022_09_17_19h24m15s"   "MM012022_05_17_13h24m33s"  
#>  [269] "MM032020_10_25_02h23m03s_1" "MM042022_09_17_05h23m33s"  
#>  [271] "MM052022_09_17_01h23m13s"   "MM082021_07_08_19h23m33s"  
#>  [273] "MM092022_09_17_19h23m18s"   "MM032020_10_25_02h22m59s"  
#>  [275] "MM042022_09_17_05h23m38s"   "MM052022_09_17_01h23m51s"  
#>  [277] "MM082021_07_08_19h23m59s"   "MM092022_09_17_19h24m01s"  
#>  [279] "MM032020_10_25_02h23m52s"   "MM042022_09_17_05h23m50s"  
#>  [281] "MM052022_09_17_01h24m16s"   "MM082021_07_08_19h23m55s"  
#>  [283] "MM092022_09_17_19h23m36s"   "MM032020_10_25_02h23m19s"  
#>  [285] "MM042022_09_17_05h22m57s"   "MM052022_09_17_01h23m25s_1"
#>  [287] "MM082021_07_08_19h23m29s"   "MM092022_09_17_19h23m37s"  
#>  [289] "MM032020_10_25_02h23m30s"   "MM042022_09_17_05h23m13s_1"
#>  [291] "MM052022_09_17_01h23m43s"   "MM082021_07_08_19h23m35s"  
#>  [293] "MM092022_09_17_19h23m53s"   "MM032020_10_25_02h23m11s_1"
#>  [295] "MM042022_09_17_05h23m03s"   "MM052022_09_17_01h23m19s"  
#>  [297] "MM082021_07_08_19h23m36s"   "MM092022_09_17_19h23m45s"  
#>  [299] "MM032020_10_25_02h23m46s_1" "MM042022_09_17_05h23m41s_1"
#>  [301] "MM052022_09_17_01h24m05s"   "MM082021_07_08_19h23m43s"  
#>  [303] "MM092022_09_17_19h23m53s_1" "MM032020_10_25_02h24m02s"  
#>  [305] "MM042022_09_17_05h23m57s"   "MM052022_09_17_01h24m03s"  
#>  [307] "MM082021_07_08_19h24m30s"   "MM092022_09_17_19h25m56s"  
#>  [309] "MM012022_05_17_13h24m09s_1" "MM032020_10_25_02h22m56s"  
#>  [311] "MM042022_09_17_05h23m32s"   "MM052022_09_17_01h25m07s"  
#>  [313] "MM082021_07_08_19h23m25s"   "MM092022_09_17_19h23m19s"  
#>  [315] "MM032020_10_25_02h23m45s"   "MM042022_09_17_05h23m21s"  
#>  [317] "MM052022_09_17_01h23m59s"   "MM082021_07_08_19h23m19s"  
#>  [319] "MM092022_09_17_19h23m29s"   "MM012022_05_17_13h23m19s_1"
#>  [321] "MM032020_10_25_03h24m52s"   "MM042022_09_17_07h23m39s"  
#>  [323] "MM052022_09_17_03h24m13s"   "MM082021_07_08_21h24m09s"  
#>  [325] "MM092022_09_17_21h24m18s"   "MM012022_05_17_15h23m55s"  
#>  [327] "MM032020_10_25_03h23m58s"   "MM042022_09_17_07h23m25s"  
#>  [329] "MM082021_07_08_21h23m51s"   "MM092022_09_17_21h23m35s"  
#>  [331] "MM012022_05_17_15h23m33s"   "MM032020_10_25_03h22m58s"  
#>  [333] "MM042022_09_17_07h23m59s"   "MM052022_09_17_03h23m47s"  
#>  [335] "MM082021_07_08_21h23m21s"   "MM092022_09_17_21h23m11s"  
#>  [337] "MM012022_05_17_15h23m58s"   "MM032020_10_25_03h24m10s"  
#>  [339] "MM042022_09_17_07h24m45s"   "MM052022_09_17_03h24m01s"  
#>  [341] "MM082021_07_08_21h24m15s"   "MM012022_05_17_15h24m39s"  
#>  [343] "MM032020_10_25_03h23m56s"   "MM042022_09_17_07h23m35s"  
#>  [345] "MM052022_09_17_03h23m37s"   "MM082021_07_08_21h23m33s"  
#>  [347] "MM092022_09_17_21h23m36s"   "MM012022_05_17_15h23m11s"  
#>  [349] "MM032020_10_25_20h24m00s"   "MM042022_09_18_00h23m49s"  
#>  [351] "MM052022_09_17_20h24m13s"   "MM082021_07_08_21h23m59s"  
#>  [353] "MM092022_09_17_21h24m17s"   "MM012022_05_17_15h23m49s"  
#>  [355] "MM032020_10_25_20h23m23s"   "MM042022_09_18_00h23m33s"  
#>  [357] "MM052022_09_17_20h24m07s"   "MM082021_07_08_21h23m42s"  
#>  [359] "MM092022_09_17_21h24m03s"   "MM012022_05_17_15h23m37s"  
#>  [361] "MM032020_10_25_20h24m16s"   "MM042022_09_18_00h23m35s"  
#>  [363] "MM052022_09_17_20h25m01s"   "MM082021_07_08_21h24m43s"  
#>  [365] "MM092022_09_17_21h25m58s"   "MM012022_05_17_15h24m28s"  
#>  [367] "MM032020_10_25_20h24m04s"   "MM042022_09_18_00h24m10s"  
#>  [369] "MM052022_09_17_20h24m35s"   "MM082021_07_08_21h23m55s"  
#>  [371] "MM012022_05_17_15h23m37s_1" "MM012022_05_17_15h24m21s"  
#>  [373] "MM032020_10_25_20h23m55s"   "MM042022_09_18_00h23m47s"  
#>  [375] "MM052022_09_17_20h23m09s"   "MM082021_07_08_21h23m53s"  
#>  [377] "MM092022_09_17_21h24m00s"   "MM012022_05_17_15h22m50s"  
#>  [379] "MM032020_10_25_20h24m22s"   "MM042022_09_18_00h24m19s"  
#>  [381] "MM052022_09_17_20h23m31s"   "MM082021_07_08_21h24m29s"  
#>  [383] "MM092022_09_17_21h24m33s"   "MM012022_05_17_15h22m47s"  
#>  [385] "MM032020_10_25_20h24m14s"   "MM042022_09_18_00h24m15s"  
#>  [387] "MM052022_09_17_20h24m52s"   "MM082021_07_08_21h24m08s"  
#>  [389] "MM092022_09_17_21h24m34s"   "MM012022_05_17_15h23m53s"  
#>  [391] "MM032020_10_25_20h23m36s"   "MM042022_09_18_00h23m09s"  
#>  [393] "MM052022_09_17_20h23m45s"   "MM082021_07_08_21h23m21s_1"
#>  [395] "MM092022_09_17_21h23m34s"   "MM012022_05_17_15h23m35s"  
#>  [397] "MM032020_10_25_20h23m20s"   "MM042022_09_18_00h23m53s"  
#>  [399] "MM052022_09_17_20h24m21s"   "MM082021_07_08_21h23m17s"  
#>  [401] "MM092022_09_17_21h23m26s"   "MM012022_05_17_15h23m23s"  
#>  [403] "MM032020_10_25_20h24m33s"   "MM042022_09_18_00h24m07s"  
#>  [405] "MM052022_09_17_20h24m45s"   "MM082021_07_08_21h23m53s_1"
#>  [407] "MM092022_09_17_21h24m06s"   "MM012022_05_17_15h23m45s"  
#>  [409] "MM032020_10_25_20h23m20s_1" "MM042022_09_18_00h23m19s"  
#>  [411] "MM052022_09_17_20h23m15s"   "MM082021_07_08_21h23m19s"  
#>  [413] "MM092022_09_17_21h24m07s"   "MM012022_05_17_15h23m44s"  
#>  [415] "MM032020_10_25_20h23m16s"   "MM042022_09_18_00h23m11s"  
#>  [417] "MM052022_09_17_20h23m51s"   "MM082021_07_08_21h23m22s"  
#>  [419] "MM092022_09_17_21h23m35s_1" "MM012022_05_17_15h24m25s"  
#>  [421] "MM032020_10_25_20h24m38s"   "MM042022_09_18_00h26m34s"  
#>  [423] "MM052022_09_17_20h24m51s"   "MM082021_07_08_21h25m22s"  
#>  [425] "MM092022_09_17_21h25m10s"   "MM012022_05_17_15h25m00s"  
#>  [427] "MM032020_10_25_20h23m56s"   "MM042022_09_18_00h25m20s"  
#>  [429] "MM052022_09_17_20h28m51s"   "MM082021_07_08_21h23m56s"  
#>  [431] "MM092022_09_17_21h23m55s"   "MM012022_05_17_15h24m33s"  
#>  [433] "MM032020_10_25_20h24m10s"   "MM042022_09_18_00h23m42s"  
#>  [435] "MM052022_09_17_20h23m47s"   "MM082021_07_08_21h23m54s"  
#>  [437] "MM092022_09_17_21h24m07s_1" "MM012022_05_17_15h23m53s_1"
#>  [439] "MM032020_10_25_20h24m06s"   "MM042022_09_18_00h23m51s"  
#>  [441] "MM052022_09_17_20h23m47s_1" "MM082021_07_08_21h24m53s"  
#>  [443] "MM092022_09_17_21h23m24s"   "MM012022_05_17_15h24m27s"  
#>  [445] "MM032020_10_25_20h23m39s"   "MM042022_09_18_00h23m35s_1"
#>  [447] "MM052022_09_17_20h22m43s"   "MM082021_07_08_21h24m07s"  
#>  [449] "MM092022_09_17_21h23m51s"   "MM012022_05_17_15h23m51s"  
#>  [451] "MM032020_10_25_20h22m56s"   "MM042022_09_18_00h22m53s"  
#>  [453] "MM052022_09_17_20h23m09s_1" "MM092022_09_17_21h23m09s"  
#>  [455] "MM012022_05_17_15h23m09s"   "MM032020_10_25_20h24m01s"  
#>  [457] "MM042022_09_18_00h24m18s"   "MM052022_09_17_20h23m21s"  
#>  [459] "MM082021_07_08_21h24m21s"   "MM092022_09_17_21h24m15s"  
#>  [461] "MM012022_05_17_15h24m29s"   "MM032020_10_25_20h23m05s"  
#>  [463] "MM042022_09_18_00h23m02s"   "MM052022_09_17_20h23m31s_1"
#>  [465] "MM082021_07_08_21h23m35s"   "MM092022_09_17_21h23m19s"  
#>  [467] "MM012022_05_17_15h23m27s"   "MM032020_10_25_20h25m14s"  
#>  [469] "MM042022_09_18_00h25m15s"   "MM052022_09_17_20h25m53s"  
#>  [471] "MM082021_07_08_21h25m15s"   "MM092022_09_17_21h25m43s"  
#>  [473] "MM012022_05_17_15h25m59s"   "MM032020_10_25_20h23m34s"  
#>  [475] "MM042022_09_18_00h23m12s"   "MM052022_09_17_20h23m49s"  
#>  [477] "MM082021_07_08_21h24m51s"   "MM092022_09_17_21h23m59s"  
#>  [479] "MM012022_05_17_15h23m39s"   "MM032020_10_25_20h25m01s"  
#>  [481] "MM042022_09_18_00h25m10s"   "MM052022_09_17_20h35m00s"  
#>  [483] "MM082021_07_08_21h34m58s"   "MM092022_09_17_21h25m26s"  
#>  [485] "MM012022_05_17_15h33m59s"   "MM032020_10_25_20h27m12s"  
#>  [487] "MM042022_09_18_00h28m24s"   "MM052022_09_17_20h27m35s"  
#>  [489] "MM082021_07_08_21h29m27s"   "MM092022_09_17_21h29m28s"  
#>  [491] "MM012022_05_17_15h29m25s"   "MM032020_10_25_20h23m03s"  
#>  [493] "MM042022_09_18_00h23m43s"   "MM052022_09_17_20h24m01s"  
#>  [495] "MM082021_07_08_21h23m35s_1" "MM092022_09_17_21h24m00s_1"
#>  [497] "MM032020_10_25_20h23m59s"   "MM042022_09_18_00h23m57s"  
#>  [499] "MM052022_09_17_20h24m33s"   "MM082021_07_08_21h24m18s"  
#>  [501] "MM092022_09_17_21h24m36s"   "MM012022_05_17_15h24m29s_1"
#>  [503] "MM032020_10_25_20h23m24s"   "MM042022_09_18_00h23m20s"  
#>  [505] "MM052022_09_17_20h23m35s"   "MM082021_07_08_21h23m25s"  
#>  [507] "MM092022_09_17_21h23m52s"   "MM012022_05_17_15h23m41s"  
#>  [509] "MM032020_10_25_20h23m19s"   "MM042022_09_18_00h22m55s"  
#>  [511] "MM052022_09_17_20h23m49s_1" "MM082021_07_08_21h22m59s"  
#>  [513] "MM092022_09_17_21h23m36s_1" "MM012022_05_17_15h23m45s_1"
#>  [515] "MM032020_10_25_23h22m59s"   "MM042022_09_18_03h23m43s"  
#>  [517] "MM052022_09_17_23h23m38s"   "MM082021_07_09_00h22m59s"  
#>  [519] "MM092022_09_18_00h23m09s"   "MM012022_05_17_18h23m43s"  
#>  [521] "MM032020_10_25_23h24m28s"   "MM042022_09_18_03h22m56s"  
#>  [523] "MM052022_09_17_23h23m19s"   "MM082021_07_09_00h23m03s"  
#>  [525] "MM092022_09_18_00h23m39s"   "MM012022_05_17_18h23m39s"  
#>  [527] "MM032020_10_25_23h23m17s"   "MM042022_09_18_03h23m07s"  
#>  [529] "MM052022_09_17_23h23m49s"   "MM082021_07_09_00h23m19s"  
#>  [531] "MM092022_09_18_00h24m23s"   "MM012022_05_17_18h24m19s"  
#>  [533] "MM032020_10_25_23h22m58s"   "MM042022_09_18_03h23m52s"  
#>  [535] "MM052022_09_17_23h23m31s"   "MM082021_07_09_00h23m03s_1"
#>  [537] "MM012022_05_17_18h23m35s"   "MM032020_10_25_23h23m42s"  
#>  [539] "MM042022_09_18_03h23m07s_1" "MM052022_09_17_23h23m47s"  
#>  [541] "MM082021_07_09_00h22m53s"   "MM092022_09_18_00h23m06s"  
#>  [543] "MM012022_05_17_18h23m21s"   "MM032020_10_25_23h24m28s_1"
#>  [545] "MM042022_09_18_03h24m33s"   "MM052022_09_17_23h24m48s"  
#>  [547] "MM082021_07_09_00h25m31s"   "MM092022_09_18_00h24m51s"  
#>  [549] "MM012022_05_17_18h25m09s"   "MM032020_10_25_23h23m04s"  
#>  [551] "MM042022_09_18_03h23m07s_2" "MM052022_09_17_23h23m48s"  
#>  [553] "MM082021_07_09_00h23m01s"   "MM092022_09_18_00h23m35s"  
#>  [555] "MM012022_05_17_18h23m37s"   "MM032020_10_25_23h23m01s"  
#>  [557] "MM042022_09_18_03h22m59s"   "MM052022_09_17_23h23m47s_1"
#>  [559] "MM082021_07_09_00h23m35s"   "MM092022_09_18_00h23m36s"  
#>  [561] "MM012022_05_17_18h23m13s"   "MM032020_10_25_23h55m30s"  
#>  [563] "MM042022_09_18_03h55m25s"   "MM052022_09_17_23h55m31s"  
#>  [565] "MM082021_07_09_00h23m41s"   "MM092022_09_18_00h23m46s"  
#>  [567] "MM012022_05_17_18h24m11s"   "MM032020_10_25_23h23m18s"  
#>  [569] "MM042022_09_18_03h23m27s"   "MM052022_09_17_23h23m52s"  
#>  [571] "MM092022_09_18_00h23m37s"   "MM012022_05_17_18h23m21s_1"
#>  [573] "MM042022_09_18_03h23m23s"   "MM052022_09_17_23h23m17s"  
#>  [575] "MM082021_07_09_00h23m41s_1" "MM092022_09_18_00h23m39s_1"
#>  [577] "MM012022_05_17_18h23m19s"   "MM032020_10_26_01h25m02s"  
#>  [579] "MM042022_09_18_05h25m13s"   "MM082021_07_09_02h24m41s"  
#>  [581] "MM092022_09_18_02h25m36s"   "MM012022_05_17_20h25m29s"  
#>  [583] "MM032020_10_26_01h23m01s"   "MM042022_09_18_05h23m00s"  
#>  [585] "MM052022_09_18_01h23m21s"   "MM082021_07_09_02h22m57s"  
#>  [587] "MM092022_09_18_02h23m27s"   "MM012022_05_17_20h22m45s"  
#>  [589] "MM032020_10_26_01h23m16s"   "MM042022_09_18_05h24m17s"  
#>  [591] "MM052022_09_18_01h23m57s"   "MM082021_07_09_02h23m13s"  
#>  [593] "MM092022_09_18_02h23m47s"   "MM012022_05_17_20h24m01s"  
#>  [595] "MM032020_10_26_01h23m27s"   "MM042022_09_18_05h23m21s"  
#>  [597] "MM052022_09_18_01h23m55s"   "MM082021_07_09_02h23m25s"  
#>  [599] "MM092022_09_18_02h23m33s"   "MM012022_05_17_20h23m12s"  
#>  [601] "MM032020_10_26_01h23m12s"   "MM042022_09_18_05h23m07s"  
#>  [603] "MM052022_09_18_01h23m18s"   "MM082021_07_09_02h23m10s"  
#>  [605] "MM092022_09_18_02h23m39s"   "MM012022_05_17_20h23m21s"  
#>  [607] "MM032020_10_26_01h23m39s"   "MM042022_09_18_05h24m00s"  
#>  [609] "MM052022_09_18_01h26m52s"   "MM082021_07_09_02h23m23s"  
#>  [611] "MM092022_09_18_02h23m39s_1" "MM012022_05_17_20h23m55s"  
#>  [613] "MM032020_10_26_01h24m55s"   "MM042022_09_18_05h24m51s"  
#>  [615] "MM052022_09_18_01h25m23s"   "MM082021_07_09_02h24m27s"  
#>  [617] "MM092022_09_18_02h24m59s"   "MM012022_05_17_20h25m35s"  
#>  [619] "MM032020_10_26_01h23m46s"   "MM042022_09_18_05h24m06s"  
#>  [621] "MM052022_09_18_01h24m19s"   "MM082021_07_09_02h24m06s"  
#>  [623] "MM092022_09_18_02h24m22s"   "MM012022_05_17_20h24m23s"  
#>  [625] "MM032020_10_26_01h24m19s"   "MM042022_09_18_05h23m55s"  
#>  [627] "MM052022_09_18_01h24m22s"   "MM082021_07_09_02h23m54s"  
#>  [629] "MM092022_09_18_02h24m13s"   "MM012022_05_17_20h24m07s"  
#>  [631] "MM032020_10_26_01h23m25s"   "MM042022_09_18_05h23m01s"  
#>  [633] "MM052022_09_18_01h23m31s"   "MM082021_07_09_02h23m45s"  
#>  [635] "MM092022_09_18_02h23m17s"   "MM012022_05_17_20h23m33s"  
#>  [637] "MM032020_10_26_01h23m09s"   "MM042022_09_18_05h22m58s"  
#>  [639] "MM052022_09_18_01h23m49s"   "MM082021_07_09_02h23m17s"  
#>  [641] "MM092022_09_18_02h23m26s"   "MM012022_05_17_20h23m46s"  
#>  [643] "MM032020_10_26_01h24m55s_1" "MM042022_09_18_05h24m58s"  
#>  [645] "MM052022_09_18_01h25m36s"   "MM082021_07_09_02h25m01s"  
#>  [647] "MM092022_09_18_02h25m12s"   "MM012022_05_17_20h25m35s_1"
#>  [649] "MM032020_10_26_01h25m05s"   "MM042022_09_18_05h27m13s"  
#>  [651] "MM052022_09_18_01h24m51s"   "MM082021_07_09_02h25m20s"  
#>  [653] "MM092022_09_18_02h24m52s"   "MM012022_05_17_20h27m19s"  
#>  [655] "MM032020_10_26_01h24m49s"   "MM042022_09_18_05h24m52s"  
#>  [657] "MM052022_09_18_01h25m05s"   "MM082021_07_09_02h25m42s"  
#>  [659] "MM092022_09_18_02h25m08s"   "MM012022_05_17_20h27m50s"  
#>  [661] "MM032020_10_26_01h22m56s"   "MM042022_09_18_05h24m55s"  
#>  [663] "MM052022_09_18_01h23m56s"   "MM082021_07_09_02h23m03s"  
#>  [665] "MM092022_09_18_02h23m51s"   "MM012022_05_17_20h23m11s"  
#>  [667] "MM032020_10_26_01h30m07s"   "MM042022_09_18_05h24m56s"  
#>  [669] "MM052022_09_18_01h23m29s"   "MM082021_07_09_02h23m17s_1"
#>  [671] "MM092022_09_18_02h23m16s"   "MM012022_05_17_20h23m33s_1"
#>  [673] "MM032020_10_26_01h22m59s"   "MM042022_09_18_05h22m43s"  
#>  [675] "MM052022_09_18_01h23m22s"   "MM082021_07_09_02h23m09s"  
#>  [677] "MM092022_09_18_02h23m04s"   "MM012022_05_17_20h22m53s"  
#>  [679] "MM032020_10_26_01h23m05s"   "MM042022_09_18_05h23m20s"  
#>  [681] "MM052022_09_18_01h23m13s"   "MM082021_07_09_02h23m13s_1"
#>  [683] "MM092022_09_18_02h23m09s"   "MM012022_05_17_20h23m43s"  
#>  [685] "MM032020_10_26_01h23m19s"   "MM042022_09_18_05h22m51s"  
#>  [687] "MM052022_09_18_01h23m17s"   "MM082021_07_09_02h23m13s_2"
#>  [689] "MM092022_09_18_02h23m33s_2" "MM012022_05_17_20h23m13s"  
#>  [691] "MM032020_10_26_01h22m59s_1" "MM042022_09_18_05h23m31s"  
#>  [693] "MM082021_07_09_02h23m29s"   "MM092022_09_18_02h23m09s_1"
#>  [695] "MM012022_05_17_20h23m47s"   "MM032020_10_26_01h23m25s_1"
#>  [697] "MM042022_09_18_05h24m57s"   "MM052022_09_18_01h24m39s"  
#>  [699] "MM082021_07_09_02h24m01s"   "MM092022_09_18_02h25m34s"  
#>  [701] "MM012022_05_17_20h24m57s"   "MM032020_10_26_01h23m14s"  
#>  [703] "MM042022_09_18_05h24m19s"   "MM052022_09_18_01h24m13s"  
#>  [705] "MM082021_07_09_02h23m53s"   "MM092022_09_18_02h24m42s"  
#>  [707] "MM012022_05_17_20h24m35s"   "MM032020_10_26_01h23m20s"  
#>  [709] "MM042022_09_18_05h23m35s"   "MM052022_09_18_01h23m54s"  
#>  [711] "MM082021_07_09_02h23m17s_2" "MM092022_09_18_02h23m53s"  
#>  [713] "MM042022_09_18_05h24m06s_1" "MM052022_09_18_01h23m41s"  
#>  [715] "MM082021_07_09_02h23m39s"   "MM012022_05_17_20h24m36s"  
#>  [717] "MM032020_10_26_01h24m24s"   "MM042022_09_18_05h24m43s"  
#>  [719] "MM052022_09_18_01h23m17s_1" "MM082021_07_09_02h24m17s"  
#>  [721] "MM092022_09_18_02h24m28s"   "MM012022_05_17_20h24m50s"  
#>  [723] "MM032020_10_26_01h23m44s"   "MM042022_09_18_05h23m30s"  
#>  [725] "MM052022_09_18_01h23m16s"   "MM082021_07_09_02h23m05s"  
#>  [727] "MM092022_09_18_02h23m01s"   "MM012022_05_17_20h23m07s"  
#>  [729] "MM032020_10_26_01h23m58s"   "MM042022_09_18_05h24m05s"  
#>  [731] "MM052022_09_18_01h24m31s"   "MM082021_07_09_02h24m05s"  
#>  [733] "MM092022_09_18_02h24m45s"   "MM012022_05_17_20h24m51s"  
#>  [735] "MM032020_10_26_01h22m54s"   "MM042022_09_18_05h23m57s"  
#>  [737] "MM052022_09_18_01h24m48s"   "MM082021_07_09_02h23m55s"  
#>  [739] "MM012022_05_17_20h24m08s"   "MM032020_10_26_01h23m38s"  
#>  [741] "MM042022_09_18_05h23m15s"   "MM052022_09_18_01h24m03s"  
#>  [743] "MM082021_07_09_02h23m51s"   "MM092022_09_18_02h24m26s"  
#>  [745] "MM012022_05_17_20h23m57s"   "MM032020_10_26_01h23m19s_1"
#>  [747] "MM042022_09_18_05h23m09s"   "MM082021_07_09_02h23m35s"  
#>  [749] "MM092022_09_18_02h23m33s_3" "MM012022_05_17_20h24m01s_1"
#>  [751] "MM032020_10_26_01h23m24s"   "MM042022_09_18_05h23m04s"  
#>  [753] "MM052022_09_18_01h23m53s"   "MM082021_07_09_02h23m02s"  
#>  [755] "MM092022_09_18_02h23m03s"   "MM012022_05_17_20h23m50s"  
#>  [757] "MM032020_10_26_01h24m25s"   "MM042022_09_18_05h24m01s"  
#>  [759] "MM052022_09_18_01h24m55s_1" "MM082021_07_09_02h24m13s"  
#>  [761] "MM092022_09_18_02h24m13s_1" "MM012022_05_17_20h25m07s"  
#>  [763] "MM032020_10_26_01h23m18s"   "MM042022_09_18_05h23m03s"  
#>  [765] "MM052022_09_18_01h23m51s"   "MM082021_07_09_02h22m55s"  
#>  [767] "MM092022_09_18_02h23m57s"   "MM012022_05_17_20h23m59s"  
#>  [769] "MM032020_10_26_01h24m32s"   "MM042022_09_18_05h23m57s_1"
#>  [771] "MM052022_09_18_01h24m45s"   "MM082021_07_09_02h22m51s"  
#>  [773] "MM092022_09_18_02h24m10s"   "MM012022_05_17_20h24m59s"  
#>  [775] "MM032020_10_26_01h23m39s_1" "MM042022_09_18_05h23m19s"  
#>  [777] "MM052022_09_18_01h24m18s"   "MM082021_07_09_02h23m25s_1"
#>  [779] "MM092022_09_18_02h24m24s"   "MM012022_05_17_20h23m27s"  
#>  [781] "MM032020_10_26_01h23m33s"   "MM042022_09_18_05h23m43s"  
#>  [783] "MM052022_09_18_01h24m29s"   "MM082021_07_09_02h23m19s"  
#>  [785] "MM092022_09_18_02h23m46s"   "MM012022_05_17_20h24m45s"  
#>  [787] "MM032020_10_26_01h23m52s"   "MM042022_09_18_05h23m03s_1"
#>  [789] "MM082021_07_09_02h23m27s"   "MM012022_05_17_20h23m19s"  
#>  [791] "MM042022_09_18_05h23m32s"   "MM052022_09_18_01h30m31s"  
#>  [793] "MM082021_07_09_02h25m33s"   "MM092022_09_18_02h26m09s"  
#>  [795] "MM012022_05_17_20h25m45s"   "MM032020_10_26_01h24m50s"  
#>  [797] "MM042022_09_18_05h23m51s"   "MM052022_09_18_01h24m11s"  
#>  [799] "MM082021_07_09_02h23m23s_1" "MM092022_09_18_02h24m22s_1"
#>  [801] "MM012022_05_17_20h24m28s"   "MM032020_10_26_01h25m33s"  
#>  [803] "MM042022_09_18_05h23m33s"   "MM052022_09_18_01h25m57s"  
#>  [805] "MM082021_07_09_02h23m45s_1" "MM092022_09_18_02h24m27s"  
#>  [807] "MM012022_05_17_20h24m39s"   "MM032020_10_26_01h28m49s"  
#>  [809] "MM042022_09_18_05h25m16s"   "MM052022_09_18_01h23m56s_1"
#>  [811] "MM082021_07_09_02h24m58s"   "MM092022_09_18_02h25m19s"  
#>  [813] "MM012022_05_17_20h25m39s"   "MM032020_10_26_01h25m32s"  
#>  [815] "MM042022_09_18_05h23m25s"   "MM052022_09_18_01h24m37s"  
#>  [817] "MM082021_07_09_02h23m21s"   "MM092022_09_18_02h23m10s"  
#>  [819] "MM012022_05_17_20h23m45s"   "MM032020_10_26_01h23m13s"  
#>  [821] "MM042022_09_18_05h23m07s_1" "MM052022_09_18_01h23m50s"  
#>  [823] "MM092022_09_18_02h23m24s"   "MM012022_05_17_20h23m39s"  
#>  [825] "MM032020_10_26_01h24m38s"   "MM042022_09_18_05h24m50s"  
#>  [827] "MM052022_09_18_01h24m23s"   "MM082021_07_09_02h24m27s_1"
#>  [829] "MM012022_05_17_20h24m05s"   "MM032020_10_26_01h24m07s"  
#>  [831] "MM042022_09_18_05h24m16s"   "MM052022_09_18_01h24m53s"  
#>  [833] "MM092022_09_18_02h24m09s"   "MM012022_05_17_20h24m39s_1"
#>  [835] "MM032020_10_26_01h25m22s"   "MM042022_09_18_05h23m24s"  
#>  [837] "MM082021_07_09_02h26m53s"   "MM092022_09_18_02h23m21s"  
#>  [839] "MM012022_05_17_20h24m37s"   "MM032020_10_26_03h24m18s"  
#>  [841] "MM042022_09_18_07h24m30s"   "MM052022_09_18_03h24m32s"  
#>  [843] "MM082021_07_09_04h23m55s"   "MM092022_09_18_04h23m17s"  
#>  [845] "MM012022_05_17_22h24m17s"   "MM032020_10_26_03h23m49s"  
#>  [847] "MM042022_09_18_07h24m06s"   "MM052022_09_18_03h24m59s"  
#>  [849] "MM082021_07_09_04h23m57s"   "MM092022_09_18_04h23m33s"  
#>  [851] "MM012022_05_17_22h23m39s"   "MM042022_09_18_07h24m23s"  
#>  [853] "MM052022_09_18_03h24m01s"   "MM082021_07_09_04h23m19s"  
#>  [855] "MM092022_09_18_04h23m15s"   "MM012022_05_17_22h23m49s"  
#>  [857] "MM032020_10_26_03h29m19s"   "MM042022_09_18_07h29m58s"  
#>  [859] "MM052022_09_18_03h24m19s"   "MM082021_07_09_04h24m15s"  
#>  [861] "MM012022_05_17_22h24m55s"   "MM032020_10_26_03h24m09s"  
#>  [863] "MM042022_09_18_07h24m25s"   "MM052022_09_18_03h25m25s"  
#>  [865] "MM082021_07_09_04h23m55s_1" "MM092022_09_18_04h23m11s"  
#>  [867] "MM012022_05_17_22h23m59s"   "MM032020_10_26_03h23m16s"  
#>  [869] "MM042022_09_18_07h23m39s"   "MM052022_09_18_03h25m09s"  
#>  [871] "MM082021_07_09_04h23m41s"   "MM092022_09_18_04h24m17s"  
#>  [873] "MM012022_05_17_22h24m27s"   "MM032020_10_26_05h23m35s"  
#>  [875] "MM042022_09_18_09h23m43s"   "MM052022_09_18_05h25m45s"  
#>  [877] "MM082021_07_09_06h25m17s"   "MM092022_09_18_06h23m13s"  
#>  [879] "MM012022_05_18_00h23m09s"   "MM032020_10_26_05h23m11s"  
#>  [881] "MM042022_09_18_09h23m11s"   "MM052022_09_18_05h25m45s_1"
#>  [883] "MM082021_07_09_06h23m29s"   "MM092022_09_18_06h23m37s"  
#>  [885] "MM012022_05_18_00h23m45s"   "MM032020_10_26_05h22m56s"  
#>  [887] "MM042022_09_18_09h23m06s"   "MM052022_09_18_05h23m13s"  
#>  [889] "MM082021_07_09_06h22m59s"   "MM092022_09_18_06h23m36s"  
#>  [891] "MM012022_05_18_00h23m08s"   "MM032020_10_26_05h23m24s"  
#>  [893] "MM042022_09_18_09h23m15s"   "MM052022_09_18_05h23m45s"  
#>  [895] "MM082021_07_09_06h23m45s"   "MM092022_09_18_06h23m32s"  
#>  [897] "MM012022_05_18_00h25m45s"   "MM032020_10_26_05h23m10s"  
#>  [899] "MM042022_09_18_09h23m38s"   "MM052022_09_18_05h23m43s"  
#>  [901] "MM082021_07_09_06h23m22s"   "MM092022_09_18_06h23m51s"  
#>  [903] "MM012022_05_18_00h23m19s"   "MM032020_10_26_05h25m06s"  
#>  [905] "MM042022_09_18_09h25m08s"   "MM052022_09_18_05h25m08s"  
#>  [907] "MM082021_07_09_06h24m45s"   "MM092022_09_18_06h25m00s"  
#>  [909] "MM012022_05_18_00h24m55s"   "MM032020_10_26_05h23m15s"  
#>  [911] "MM042022_09_18_09h23m16s"   "MM082021_07_09_06h23m18s"  
#>  [913] "MM012022_05_18_00h23m59s"   "MM032020_10_26_05h23m12s"  
#>  [915] "MM042022_09_18_09h23m07s"   "MM052022_09_18_05h23m15s"  
#>  [917] "MM082021_07_09_06h23m27s"   "MM012022_05_18_00h23m47s"  
#>  [919] "MM032020_10_26_05h23m37s"   "MM042022_09_18_09h23m57s"  
#>  [921] "MM052022_09_18_05h23m57s"   "MM082021_07_09_06h23m47s"  
#>  [923] "MM092022_09_18_06h24m21s"   "MM012022_05_18_00h24m17s"  
#>  [925] "MM032020_10_26_05h23m46s"   "MM042022_09_18_09h23m08s"  
#>  [927] "MM052022_09_18_05h23m55s"   "MM082021_07_09_06h23m23s"  
#>  [929] "MM092022_09_18_06h23m31s"   "MM012022_05_18_00h23m29s"  
#>  [931] "MM032020_10_26_05h22m59s"   "MM042022_09_18_09h23m11s_1"
#>  [933] "MM052022_09_18_05h24m21s"   "MM082021_07_09_06h23m56s"  
#>  [935] "MM092022_09_18_06h24m18s"   "MM012022_05_18_00h24m15s"  
#>  [937] "MM032020_10_26_05h22m54s"   "MM042022_09_18_09h22m53s"  
#>  [939] "MM052022_09_18_05h24m25s"   "MM082021_07_09_06h23m13s"  
#>  [941] "MM092022_09_18_06h24m06s"   "MM012022_05_18_00h23m11s"  
#>  [943] "MM032020_10_26_05h24m21s"   "MM042022_09_18_09h24m14s"  
#>  [945] "MM052022_09_18_05h24m47s"   "MM082021_07_09_06h24m10s"  
#>  [947] "MM092022_09_18_06h24m35s"   "MM012022_05_18_00h24m26s"  
#>  [949] "MM032020_10_26_05h23m36s"   "MM042022_09_18_09h23m41s"  
#>  [951] "MM052022_09_18_05h24m25s_1" "MM082021_07_09_06h23m52s"  
#>  [953] "MM092022_09_18_06h23m52s"   "MM012022_05_18_00h23m57s"  
#>  [955] "MM032020_10_26_05h22m55s"   "MM042022_09_18_09h23m25s"  
#>  [957] "MM052022_09_18_05h23m05s"   "MM082021_07_09_06h22m59s_1"
#>  [959] "MM092022_09_18_06h23m22s"   "MM012022_05_18_00h23m24s"  
#>  [961] "MM032020_10_26_05h24m08s"   "MM042022_09_18_09h24m13s"  
#>  [963] "MM052022_09_18_05h23m47s"   "MM082021_07_09_06h23m39s"  
#>  [965] "MM092022_09_18_06h24m24s"   "MM012022_05_18_00h24m18s"  
#>  [967] "MM032020_10_26_05h25m51s"   "MM042022_09_18_09h26m13s"  
#>  [969] "MM052022_09_18_05h26m43s"   "MM082021_07_09_06h25m57s"  
#>  [971] "MM092022_09_18_06h26m28s"   "MM012022_05_18_00h26m21s"  
#>  [973] "MM032020_10_26_05h23m40s"   "MM042022_09_18_09h23m00s"  
#>  [975] "MM052022_09_18_05h23m47s_1" "MM082021_07_09_06h23m11s"  
#>  [977] "MM092022_09_18_06h23m20s"   "MM012022_05_18_00h23m51s"  
#>  [979] "MM032020_10_26_05h34m38s"   "MM042022_09_18_09h23m53s"  
#>  [981] "MM052022_09_18_05h24m35s"   "MM082021_07_09_06h24m01s"  
#>  [983] "MM092022_09_18_06h24m43s"   "MM012022_05_18_00h24m41s"  
#>  [985] "MM042022_09_18_09h23m44s"   "MM052022_09_18_05h24m09s"  
#>  [987] "MM082021_07_09_06h23m38s"   "MM092022_09_18_06h24m29s"  
#>  [989] "MM012022_05_18_00h24m00s"   "MM032020_10_26_05h26m02s"  
#>  [991] "MM042022_09_18_09h26m03s"   "MM052022_09_18_05h26m39s"  
#>  [993] "MM082021_07_09_06h26m05s"   "MM092022_09_18_06h26m20s"  
#>  [995] "MM012022_05_18_00h26m31s"   "MM032020_10_26_05h24m06s"  
#>  [997] "MM042022_09_18_09h26m19s"   "MM052022_09_18_05h24m25s_2"
#>  [999] "MM092022_09_18_06h24m51s"   "MM032020_10_26_05h23m32s"  
#> [1001] "MM042022_09_18_09h23m45s"   "MM052022_09_18_05h24m11s"  
#> [1003] "MM082021_07_09_06h23m51s"   "MM092022_09_18_06h25m49s"  
#> [1005] "MM012022_05_18_00h24m16s"   "MM032020_10_26_05h22m54s_1"
#> [1007] "MM042022_09_18_09h23m14s"   "MM052022_09_18_05h24m03s"  
#> [1009] "MM082021_07_09_06h23m15s"   "MM092022_09_18_06h24m01s"  
#> [1011] "MM012022_05_18_00h23m19s_1" "MM032020_10_26_05h23m08s"  
#> [1013] "MM042022_09_18_09h23m33s"   "MM082021_07_09_06h24m18s"  
#> [1015] "MM092022_09_18_06h23m09s"   "MM012022_05_18_00h23m13s"  
#> [1017] "MM032020_10_26_05h23m03s"   "MM042022_09_18_09h23m09s"  
#> [1019] "MM052022_09_18_05h23m55s_1" "MM082021_07_09_06h23m11s_1"
#> [1021] "MM092022_09_18_06h23m11s"   "MM012022_05_18_00h23m13s_1"
#> [1023] "MM032020_10_26_05h23m27s"   "MM042022_09_18_09h23m05s"  
#> [1025] "MM052022_09_18_05h24m22s"   "MM082021_07_09_06h23m17s"  
#> [1027] "MM092022_09_18_06h23m29s"   "MM012022_05_18_00h23m25s"  
#> [1029] "MM032020_10_26_05h23m35s_1" "MM042022_09_18_09h23m41s_1"
#> [1031] "MM052022_09_18_05h23m20s"   "MM082021_07_09_06h23m43s"  
#> [1033] "MM092022_09_18_06h23m39s"   "MM012022_05_18_00h23m15s"  
#> [1035] "MM032020_10_26_05h23m28s"   "MM042022_09_18_09h23m29s"  
#> [1037] "MM052022_09_18_05h23m55s_2" "MM082021_07_09_06h23m37s"  
#> [1039] "MM092022_09_18_06h23m45s"   "MM012022_05_18_00h23m57s_1"
#> [1041] "MM032020_10_26_05h22m54s_2" "MM042022_09_18_09h23m06s_1"
#> [1043] "MM052022_09_18_05h23m17s"   "MM082021_07_09_06h22m55s"  
#> [1045] "MM092022_09_18_06h23m05s"   "MM012022_05_18_00h23m07s"  
#> [1047] "MM052022_09_18_05h24m12s"   "MM082021_07_09_06h24m59s"  
#> [1049] "MM092022_09_18_06h25m51s"   "MM012022_05_18_00h25m57s"  
#> [1051] "MM032020_10_26_05h23m21s"   "MM042022_09_18_09h25m46s"  
#> [1053] "MM052022_09_18_05h24m24s"   "MM082021_07_09_06h23m59s"  
#> [1055] "MM092022_09_18_06h25m44s"   "MM012022_05_18_00h24m50s"  
#> [1057] "MM032020_10_26_05h23m30s"   "MM042022_09_18_09h23m29s_1"
#> [1059] "MM052022_09_18_05h23m39s_1" "MM082021_07_09_06h23m39s_1"
#> [1061] "MM092022_09_18_06h23m13s_1" "MM012022_05_18_00h24m53s"  
#> [1063] "MM032020_10_26_05h24m35s"   "MM042022_09_18_09h24m35s"  
#> [1065] "MM052022_09_18_05h24m38s"   "MM082021_07_09_06h25m02s"  
#> [1067] "MM092022_09_18_06h23m18s"   "MM012022_05_18_00h24m03s"  
#> [1069] "MM032020_10_26_05h23m17s"   "MM042022_09_18_09h24m42s"  
#> [1071] "MM052022_09_18_05h23m16s"   "MM082021_07_09_06h24m09s"  
#> [1073] "MM092022_09_18_06h23m43s"   "MM012022_05_18_00h24m13s"  
#> [1075] "MM032020_10_26_05h24m02s"   "MM042022_09_18_09h23m48s"  
#> [1077] "MM052022_09_18_05h25m15s"   "MM082021_07_09_06h25m13s"  
#> [1079] "MM092022_09_18_06h23m47s"   "MM012022_05_18_00h24m02s"  
#> [1081] "MM032020_10_26_05h26m32s"   "MM042022_09_18_09h25m09s"  
#> [1083] "MM052022_09_18_05h25m11s"   "MM082021_07_09_06h25m04s"  
#> [1085] "MM092022_09_18_06h25m09s"   "MM032020_10_26_05h23m37s_1"
#> [1087] "MM042022_09_18_09h23m02s"   "MM052022_09_18_05h23m49s"  
#> [1089] "MM082021_07_09_06h23m05s"   "MM092022_09_18_06h23m52s_1"
#> [1091] "MM032020_10_26_05h26m44s"   "MM042022_09_18_09h26m45s"  
#> [1093] "MM052022_09_18_05h27m01s"   "MM082021_07_09_06h26m58s"  
#> [1095] "MM092022_09_18_06h27m09s"   "MM012022_05_18_00h22m59s"  
#> [1097] "MM032020_10_26_05h23m28s_1" "MM042022_09_18_09h22m59s"  
#> [1099] "MM052022_09_18_05h23m48s"   "MM082021_07_09_06h23m09s"  
#> [1101] "MM092022_09_18_06h23m28s"   "MM012022_05_18_00h23m25s_1"
#> [1103] "MM032020_10_26_05h24m53s"   "MM042022_09_18_09h24m50s"  
#> [1105] "MM052022_09_18_05h25m09s"   "MM082021_07_09_06h24m45s_1"
#> [1107] "MM092022_09_18_06h25m09s_1" "MM012022_05_18_00h24m54s"  
#> [1109] "MM032020_10_26_05h24m16s"   "MM042022_09_18_09h24m19s"  
#> [1111] "MM052022_09_18_05h23m59s"   "MM082021_07_09_06h23m47s_1"
#> [1113] "MM092022_09_18_06h23m44s"   "MM012022_05_18_00h23m55s"  
#> [1115] "MM032020_10_26_05h25m03s"   "MM052022_09_18_05h25m14s"  
#> [1117] "MM082021_07_09_06h24m35s"   "MM092022_09_18_06h24m43s_1"
#> [1119] "MM012022_05_18_00h25m07s"   "MM032020_10_26_05h23m30s_1"
#> [1121] "MM042022_09_18_09h23m39s"   "MM052022_09_18_05h23m35s"  
#> [1123] "MM082021_07_09_06h23m46s"   "MM092022_09_18_06h23m34s"  
#> [1125] "MM012022_05_18_00h23m35s"   "MM032020_10_26_06h24m05s"  
#> [1127] "MM042022_09_18_10h24m05s"   "MM052022_09_18_06h24m30s"  
#> [1129] "MM082021_07_09_07h24m17s"   "MM092022_09_18_07h24m08s"  
#> [1131] "MM012022_05_18_01h24m23s"   "MM032020_10_26_06h23m11s"  
#> [1133] "MM042022_09_18_10h23m07s"   "MM052022_09_18_06h23m59s"  
#> [1135] "MM082021_07_09_07h23m48s"   "MM092022_09_18_07h23m44s"  
#> [1137] "MM012022_05_18_01h23m33s"   "MM032020_10_26_06h23m30s"  
#> [1139] "MM042022_09_18_10h23m07s_1" "MM052022_09_18_06h24m04s"  
#> [1141] "MM082021_07_09_07h23m35s"   "MM092022_09_18_07h23m46s"  
#> [1143] "MM012022_05_18_01h23m43s"   "MM032020_10_26_06h23m30s_1"
#> [1145] "MM042022_09_18_10h23m44s"   "MM052022_09_18_06h24m03s"  
#> [1147] "MM082021_07_09_07h25m27s"   "MM012022_05_18_01h24m00s"  
#> [1149] "MM032020_10_26_06h23m29s"   "MM042022_09_18_10h23m31s"  
#> [1151] "MM052022_09_18_06h23m55s"   "MM082021_07_09_07h23m17s"  
#> [1153] "MM092022_09_18_07h23m54s"   "MM012022_05_18_01h23m45s"  
#> [1155] "MM032020_10_26_06h23m50s"   "MM042022_09_18_10h23m29s"  
#> [1157] "MM052022_09_18_06h26m23s"   "MM082021_07_09_07h23m00s"  
#> [1159] "MM092022_09_18_07h23m22s"   "MM012022_05_18_01h23m47s"  
#> [1161] "MM032020_10_26_06h23m15s"   "MM042022_09_18_10h23m01s"  
#> [1163] "MM052022_09_18_06h23m21s"   "MM082021_07_09_07h23m55s"  
#> [1165] "MM092022_09_18_07h23m21s"   "MM012022_05_18_01h23m03s"  
#> [1167] "MM032020_10_26_06h23m08s"   "MM042022_09_18_10h23m03s"  
#> [1169] "MM052022_09_18_06h23m34s"   "MM082021_07_09_07h23m56s"  
#> [1171] "MM092022_09_18_07h23m57s"   "MM012022_05_18_01h23m05s"  
#> [1173] "MM032020_10_26_06h22m46s"   "MM042022_09_18_10h23m40s"  
#> [1175] "MM052022_09_18_06h24m13s"   "MM082021_07_09_07h23m35s_1"
#> [1177] "MM092022_09_18_07h23m22s_1" "MM012022_05_18_01h23m53s"  
#> [1179] "MM032020_10_26_06h22m51s"   "MM052022_09_18_06h22m45s"  
#> [1181] "MM082021_07_09_07h23m01s"   "MM092022_09_18_07h25m38s"  
#> [1183] "MM012022_05_18_01h24m22s"   "MM032020_10_26_07h25m06s"  
#> [1185] "MM042022_09_18_11h25m19s"   "MM052022_09_18_07h25m47s"  
#> [1187] "MM082021_07_09_08h25m38s"   "MM092022_09_18_08h25m21s"  
#> [1189] "MM012022_05_18_02h25m19s"   "MM032020_10_26_07h25m02s"  
#> [1191] "MM042022_09_18_11h25m21s"   "MM052022_09_18_07h25m06s"  
#> [1193] "MM082021_07_09_08h25m27s"   "MM092022_09_18_08h25m25s"  
#> [1195] "MM012022_05_18_02h25m11s"   "MM032020_10_26_07h23m27s"  
#> [1197] "MM042022_09_18_11h23m25s"   "MM052022_09_18_07h23m45s"  
#> [1199] "MM082021_07_09_08h23m01s"   "MM092022_09_18_08h23m10s"  
#> [1201] "MM012022_05_18_02h22m55s"   "MM032020_10_26_07h23m22s"  
#> [1203] "MM042022_09_18_11h23m43s"   "MM052022_09_18_07h23m55s"  
#> [1205] "MM082021_07_09_08h23m53s"   "MM092022_09_18_08h23m10s_1"
#> [1207] "MM012022_05_18_02h23m48s"   "MM032020_10_26_07h24m42s"  
#> [1209] "MM042022_09_18_11h24m34s"   "MM052022_09_18_07h25m17s"  
#> [1211] "MM082021_07_09_08h25m07s"   "MM092022_09_18_08h25m10s"  
#> [1213] "MM012022_05_18_02h26m20s"   "MM032020_10_26_07h24m01s"  
#> [1215] "MM042022_09_18_11h23m19s"   "MM052022_09_18_07h24m21s"  
#> [1217] "MM082021_07_09_08h23m42s"   "MM092022_09_18_08h24m52s"  
#> [1219] "MM012022_05_18_02h24m49s"   "MM032020_10_26_07h23m01s"  
#> [1221] "MM042022_09_18_11h23m11s"   "MM052022_09_18_07h23m35s"  
#> [1223] "MM082021_07_09_08h23m16s"   "MM092022_09_18_08h23m31s"  
#> [1225] "MM012022_05_18_02h23m43s"   "MM032020_10_26_07h23m32s"  
#> [1227] "MM042022_09_18_11h23m53s"   "MM052022_09_18_07h23m39s"  
#> [1229] "MM082021_07_09_08h23m25s"   "MM012022_05_18_02h24m15s"  
#> [1231] "MM042022_09_18_11h23m11s_1" "MM052022_09_18_07h37m17s"  
#> [1233] "MM082021_07_09_08h37m21s"   "MM092022_09_18_08h37m31s"  
#> [1235] "MM032020_10_26_07h22m59s"   "MM042022_09_18_11h23m45s"  
#> [1237] "MM052022_09_18_07h23m30s"   "MM082021_07_09_08h23m41s"  
#> [1239] "MM092022_09_18_08h23m34s"   "MM012022_05_18_02h23m39s"  
#> [1241] "MM032020_10_26_07h23m23s"   "MM042022_09_18_11h23m45s_1"
#> [1243] "MM052022_09_18_07h23m55s_1" "MM082021_07_09_08h23m27s"  
#> [1245] "MM092022_09_18_08h24m09s"   "MM012022_05_18_02h23m33s"  
#> [1247] "MM032020_10_26_07h22m58s"   "MM042022_09_18_11h23m07s"  
#> [1249] "MM052022_09_18_07h23m25s"   "MM082021_07_09_08h23m15s"  
#> [1251] "MM092022_09_18_08h23m40s"   "MM012022_05_18_02h23m11s"  
#> [1253] "MM032020_10_26_07h22m54s"   "MM042022_09_18_11h22m48s"  
#> [1255] "MM052022_09_18_07h23m41s"   "MM082021_07_09_08h22m41s"  
#> [1257] "MM092022_09_18_08h27m05s"   "MM012022_05_18_02h23m03s"  
#> [1259] "MM032020_10_26_07h23m17s"   "MM042022_09_18_11h24m07s"  
#> [1261] "MM052022_09_18_07h23m48s"   "MM082021_07_09_08h22m59s"  
#> [1263] "MM092022_09_18_08h22m49s"   "MM012022_05_18_02h23m36s"  
#> [1265] "MM032020_10_26_07h23m31s"   "MM042022_09_18_11h23m29s"  
#> [1267] "MM052022_09_18_07h23m43s"   "MM082021_07_09_08h23m35s"  
#> [1269] "MM092022_09_18_08h23m43s"   "MM012022_05_18_02h23m21s"  
#> [1271] "MM032020_10_26_07h23m01s_1" "MM042022_09_18_11h23m04s"  
#> [1273] "MM052022_09_18_07h23m04s"   "MM082021_07_09_08h23m06s"  
#> [1275] "MM012022_05_18_02h23m24s"   "MM032020_10_26_07h22m49s"  
#> [1277] "MM042022_09_18_11h22m46s"   "MM052022_09_18_07h23m01s"  
#> [1279] "MM082021_07_09_08h22m53s"   "MM092022_09_18_08h23m02s"  
#> [1281] "MM012022_05_18_02h23m01s"   "MM032020_10_26_07h23m05s"  
#> [1283] "MM042022_09_18_11h23m01s"   "MM052022_09_18_07h23m11s"  
#> [1285] "MM082021_07_09_08h23m07s"   "MM092022_09_18_08h23m03s"  
#> [1287] "MM012022_05_18_02h23m15s"   "MM032020_10_26_07h22m55s"  
#> [1289] "MM042022_09_18_11h23m03s"   "MM052022_09_18_07h23m47s"  
#> [1291] "MM082021_07_09_08h22m59s_1" "MM092022_09_18_08h23m37s"  
#> [1293] "MM012022_05_18_02h23m15s_1" "MM042022_09_18_11h23m25s_1"
#> [1295] "MM052022_09_18_07h24m11s"   "MM082021_07_09_08h23m08s"  
#> [1297] "MM012022_05_18_02h23m19s"   "MM032020_10_26_07h23m05s_1"
#> [1299] "MM042022_09_18_11h23m08s"   "MM052022_09_18_07h23m31s"  
#> [1301] "MM082021_07_09_08h23m09s"   "MM092022_09_18_08h23m22s"  
#> [1303] "MM012022_05_18_02h23m31s"   "MM032020_10_26_07h23m26s"  
#> [1305] "MM042022_09_18_11h23m51s"   "MM052022_09_18_07h24m30s"  
#> [1307] "MM082021_07_09_08h24m47s"   "MM092022_09_18_08h22m50s"  
#> [1309] "MM012022_05_18_02h24m07s"   "MM032020_10_26_07h23m17s_1"
#> [1311] "MM042022_09_18_11h23m21s"   "MM052022_09_18_07h23m47s_1"
#> [1313] "MM082021_07_09_08h23m07s_1" "MM092022_09_18_08h23m23s"  
#> [1315] "MM012022_05_18_02h23m17s"   "MM032020_10_26_07h22m56s"  
#> [1317] "MM042022_09_18_11h23m03s_1" "MM052022_09_18_07h23m15s"  
#> [1319] "MM082021_07_09_08h23m07s_2" "MM092022_09_18_08h23m17s"  
#> [1321] "MM012022_05_18_02h23m57s"   "MM032020_10_26_08h23m36s"  
#> [1323] "MM042022_09_18_12h23m33s"   "MM052022_09_18_08h23m59s"  
#> [1325] "MM082021_07_09_09h24m54s"   "MM092022_09_18_09h23m46s"  
#> [1327] "MM012022_05_18_03h24m01s"   "MM032020_10_26_08h22m54s"  
#> [1329] "MM042022_09_18_12h24m57s"   "MM052022_09_18_08h25m03s"  
#> [1331] "MM082021_07_09_09h26m17s"   "MM012022_05_18_03h24m45s"  
#> [1333] "MM032020_10_26_08h23m32s"   "MM042022_09_18_12h23m47s"  
#> [1335] "MM052022_09_18_08h23m45s"   "MM082021_07_09_09h23m46s"  
#> [1337] "MM092022_09_18_09h24m03s"   "MM012022_05_18_03h23m50s"  
#> [1339] "MM032020_10_26_08h24m36s"   "MM042022_09_18_12h24m17s"  
#> [1341] "MM052022_09_18_08h24m45s"   "MM082021_07_09_09h24m33s"  
#> [1343] "MM092022_09_18_09h24m42s"   "MM012022_05_18_03h24m55s"  
#> [1345] "MM032020_10_26_08h23m24s"   "MM042022_09_18_12h23m14s"  
#> [1347] "MM052022_09_18_08h24m45s_1" "MM082021_07_09_09h24m55s"  
#> [1349] "MM012022_05_18_03h24m27s"   "MM032020_10_26_08h24m17s"  
#> [1351] "MM042022_09_18_12h24m32s"   "MM052022_09_18_08h23m27s"  
#> [1353] "MM082021_07_09_09h24m25s"   "MM092022_09_18_09h23m47s"  
#> [1355] "MM012022_05_18_03h24m34s"   "MM032020_10_26_08h23m04s"  
#> [1357] "MM042022_09_18_12h23m11s"   "MM052022_09_18_08h23m44s"  
#> [1359] "MM082021_07_09_09h23m19s"   "MM092022_09_18_09h23m05s"  
#> [1361] "MM012022_05_18_03h23m25s"   "MM032020_10_26_10h27m19s"  
#> [1363] "MM042022_09_18_14h27m07s"   "MM052022_09_18_10h27m35s"  
#> [1365] "MM082021_07_09_11h27m35s"   "MM092022_09_18_11h27m29s"  
#> [1367] "MM012022_05_18_05h27m29s"   "MM032020_10_26_10h23m08s"  
#> [1369] "MM042022_09_18_14h23m07s"   "MM052022_09_18_10h23m26s"  
#> [1371] "MM082021_07_09_11h23m25s"   "MM092022_09_18_11h23m18s"  
#> [1373] "MM012022_05_18_05h27m33s"   "MM032020_10_26_10h23m24s"  
#> [1375] "MM042022_09_18_14h23m37s"   "MM052022_09_18_10h23m13s"  
#> [1377] "MM082021_07_09_11h23m09s"   "MM092022_09_18_11h23m32s"  
#> [1379] "MM032020_10_26_10h23m13s"   "MM042022_09_18_14h23m18s"  
#> [1381] "MM052022_09_18_10h23m36s"   "MM082021_07_09_11h23m07s"  
#> [1383] "MM092022_09_18_11h23m18s_1" "MM012022_05_18_05h23m19s"  
#> [1385] "MM032020_10_26_10h23m19s"   "MM042022_09_18_14h24m08s"  
#> [1387] "MM052022_09_18_10h23m39s"   "MM082021_07_09_11h23m37s"  
#> [1389] "MM092022_09_18_11h23m33s"   "MM012022_05_18_05h23m33s"  

mm_csv_unfiltered <- import_mitemap(
  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  file_name_column = "File (mite ID)", verbose = FALSE, clean=FALSE
)
dim(mm_csv_unfiltered)
#> [1] 76032    35
mm_csv_unfiltered$File_name |> unique() |> length()
#> [1] 53

mm_xlsx <- import_mitemap(
  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  format_metadata = "xlsx", verbose = FALSE
)
#> Warning: Expecting numeric in I1231 / R1231C9: got 'NA'
#> Warning: Expecting numeric in L1231 / R1231C12: got 'NA'
#> Warning: Expecting numeric in I1237 / R1237C9: got 'NA'
#> Warning: Expecting numeric in L1237 / R1237C12: got 'NA'
#> Warning: Expecting numeric in I1514 / R1514C9: got 'NA'
#> Warning: Expecting numeric in L1514 / R1514C12: got 'NA'
#> Warning: Expecting numeric in I1515 / R1515C9: got 'NA'
#> Warning: Expecting numeric in L1515 / R1515C12: got 'NA'
#> Warning: Expecting numeric in I1516 / R1516C9: got 'NA'
#> Warning: Expecting numeric in L1516 / R1516C12: got 'NA'
#> Warning: Expecting numeric in I1517 / R1517C9: got 'NA'
#> Warning: Expecting numeric in L1517 / R1517C12: got 'NA'
#> Warning: Expecting numeric in I1518 / R1518C9: got 'NA'
#> Warning: Expecting numeric in L1518 / R1518C12: got 'NA'
#> Warning: Expecting numeric in I1519 / R1519C9: got 'NA'
#> Warning: Expecting numeric in L1519 / R1519C12: got 'NA'
#> New names:
#> • `MiteID` -> `MiteID...1`
#> • `MiteID` -> `MiteID...4`

dim(mm_xlsx)
#> [1] 70828    35

# mm_xlsx$Date <-
#   as.POSIXct(strptime(mm_xlsx$Date, "%d/%m/%Y"))

mm_wo_metadata <- import_mitemap(
  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  format_metadata = NULL, verbose = FALSE
)

dim(mm_wo_metadata)
#> [1] 72257    16